kernel: cap_delete: clean up delete list handling.
This commit ensures that the delete list is always in a consistent state
when a function that could potentially add elements to it is called.

This is necessary, because when there is one CNode remaining in the
delete list, this CNode needs to be shuffled into the clear list by
caps_delete_last(). For this to happen, the CNode must not be in the
delete list anymore. Previously we just cached the next element of the
delete list, cleared the CNode's delete list pointer, but didn't fully
remove the cte from the delete list.

However, there is the possibility that caps_delete_last() will also
insert capabilities in that CNode into the delete list.  Previously
these insertions would happen on the inconsistent delete list, and would
be lost after caps_delete_last() completes without errors, as
caps_delete_step() used a cached version of the delete list's next
element (i.e. NULL in this scenario) to restore the delete list to a
consistent state, thereby losing the elements just added to the delete
list by caps_delete_last().

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>