hake: Rewrite libraryOs selection to be more robust.
This commit replaces the String libraryOs in Args.Args with a (Maybe
Args) and the rule macros for building applications and libraries
respect the fields addCFlags, addCxxFlags, omitCFlags, omitCxxFlags, and
addLinkFlags of the provided Args.libraryOs.
The Args.target of the Args.libraryOs of the application macro is used
to select the right library os archive for the linker.
The Library OS Hakefile itself needs to set the library's Args.libraryOs
to an empty Args because a Nothing Args.libraryOs will be replaced with
an architecture-specific default library OS as defined in Config.hs.

All available library OS implementations need to have a (Maybe Args) in
Config.hs which can then be specified in application and library
Hakefiles.

This commit also introduces more libbarrelfish variants for all
combinations of {pmap_array, pmap_ll} x {slot_alloc'd mappings, mapping
cnodes}, for architectures that support different pmap implementations,
i.e. x86_64, aarch64 and k1om.

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>